+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## 3.0
 
 ### NEXT
+* HMAC Support
+* AES Support
+    * CBC
+    * CBC-HMAC
+    * GCM
+
 
 ### 3.11.1 (Supreme 0.6.1)
 * Fix `CoseSigned` JSON serialization